pyproject.toml
[tool.poetry]
name = "openapi-collector"
version = "0.1.3"
description = ""
authors = ["Daniel Middlecote <dan@dlmiddlecote.io>"]

[tool.poetry.dependencies]
python = "^3.7"
pykube-ng = {version = "^19.9", extras = ["gcp"]}
requests = "^2.22"
flask = "^1.1"
[tool.poetry.dev-dependencies]
flake8 = "^3.7"
pytest = "^5.1"
pytest-cov = "^2.7"
coveralls = "^1.8"
coverage = "^4.5"
pylint = "^2.3"
pytest-kind = "^19.9"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
